requirements.txt
annotated-types==0.7.0
anyio==4.12.1
attrs==25.3.0
audioop-lts==0.2.2
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
comtypes==1.4.15
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.1
keyboard==0.13.5
MarkupSafe==3.0.2
MouseInfo==0.1.3
numpy==2.2.4
openai==2.21.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
psycopg2==2.9.10
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
PyGetWindow==0.0.9
PyMsgBox==2.0.1
PyMuPDF==1.25.5
pyperclip==1.11.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyttsx3==2.99
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
pywin32==311
requests==2.32.3
scikit-learn==1.7.1
scipy==1.16.1
selenium==4.31.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.14.5
standard-aifc==3.13.0
standard-chunk==3.13.0
threadpoolctl==3.6.0
tqdm==4.67.3
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.35.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wikipedia==1.4.0
wsproto==1.2.0
